convert size_t to int

+++ src/tree/updater_gpu.cu
@@ -596,7 +596,7 @@ class GPUMaker : public TreeUpdater {
     try {
       // build tree
       for (size_t i = 0; i < trees.size(); ++i) {
-        printf("tree_construction_id=%d\n", i);
+        printf("tree_construction_id=%d\n", static_cast<int>(i));
         UpdateTree(gpair, dmat, trees[i]);
       }
     } catch (const std::exception& e) {
